version="logparser.py v0.2.0"
Accompanying "akparser3.py" module is self declared to be
version 0.2.0 as well.

ipDic continues to be a multidepth dictionary. It's order is as follows:
IP address, file_type, file_name, ....
... depends on file_type:
  ... if it is  a black or white file (not an input/log file:
        The value is an integer used as a counter.
  --- if it's a log file:
        The value is an instance of the IP_Class which provides its own
        methods of entering and retrieving data.

A typical test command:
 ./logparser.py3 -b black.sh -w white.sh -i f2.log \
 -i nonexistant_file -i a.log -o out

Output goes to STDOUT unless a file is specified using the -o option.
Currently there are still debugging statements in the code and they are
sent to STDOUT.

TODO: I would like to change the code so that the main ouput is in order of
frequency instead of ordered by IP number.