version="logparser.py v0.2.1"
Accompanying "akparser3.py" module is self declared to be
version 0.2.1 as well.

Resolved problem that counts didn't match.
Moved a lot of the code out of the main body and
into methods and procedures.

Typical test commands:
 ./logparser.py3 -b black.sh -w white.sh -i f2.log -i a.log >out
 ./logparser.py3 -b black.sh -w white.sh -i f2.log -i a.log -o out
 ./logparser.py3 -b black.sh -i nonexist -w white.sh -i f2.log -i a.log -o out
 ./logparser.py3 -vrfb black.sh -i nonexist -w white.sh -i f2.log -i a.log -o out

Output goes to STDOUT unless a file is specified using the -o option.
Debugging statements which will eventually be removed are sent to STDOUT.
DEBUGGING REPORT not really used.  Will probably be eliminated.