version="logparser3.py v0.2.5"
The result of a clean up in preparation for planned sharing of this code
with Glen Jarvis who has indicated he might be willing to look at it and
perhaps even collaborate.

Previously I've been doing everything on the master branch.  After this
commit, the plan is to do things the 'gitflow' way.

Accompanying "akparser3.py" module is self declared to be
version 0.2.5 as well.

Fixed ability to use stdin rather than specifying --input files.

Typical test commands:
 ./logparser3.py -b black.sh -w white.sh -i f2.log -i a.log >out
 ./logparser3.py -b black.sh -w white.sh -i f2.log -i a.log -o out
 ./logparser3.py -b black.sh -i nonexist -w white.sh -i f2.log -i a.log \
-o out
 ./logparser3.py -vrfb black.sh -i nonexist -w white.sh -i f2.log \
-i a.log -o out
 ./logparser3.py -vrrfl ~/Logs -w white.sh -o out

Output goes to STDOUT unless a file is specified using the -o option.

Debugging statements which will eventually be removed are sent to STDOUT.
DEBUGGING REPORT not really used.  Will probably be eliminated.